policy_agent_design/requirements.txt
# 부동산 정책 분석 Agent 시스템 패키지 의존성
# Python 3.8+ 권장

# PDF 처리
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10

# LangChain & LangGraph
langchain==0.1.16
langgraph==0.0.28
langchain-openai==0.1.0
langchain-community==0.0.34

# 벡터 DB & 임베딩
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
chromadb==0.4.22  # 대안 벡터 DB

# OpenAI
openai==1.12.0
tiktoken==0.5.2

# 데이터 처리
numpy==1.24.4
pandas==2.1.4
pyyaml==6.0.1

# 유틸리티
python-dotenv==1.0.0
tenacity==8.2.3
tqdm==4.66.1
colorama==0.4.6

# 웹 인터페이스 (선택사항)
# streamlit==1.31.0
# gradio==4.19.0

# 모니터링 (선택사항)
# wandb==0.16.3
# loguru==0.7.2

# 테스트 (선택사항)
# pytest==8.0.0
# pytest-asyncio==0.23.3
